+++ test/unit/media-session.test.ts
@@ -8,12 +8,15 @@ class FakePeerConnection extends EventTarget {
   addTrack = jest.fn();
   getSenders = jest.fn();
   getTransceivers = jest.fn();
+  close = jest.fn();
 }
 
 class FakeParent extends EventEmitter {
   createPeerConnection () {
     return new FakePeerConnection();
-  }
+  };
+  signal = jest.fn();
+  forgetSession = jest.fn();
 }
 
 describe('GenesysCloudMediaSession', () => {
@@ -28,6 +31,81 @@ describe('GenesysCloudMediaSession', () => {
       });
 
       expect(session['statsGatherer']).toBeFalsy();
+    });
+  });
+
+ describe('end', () => {
+    afterEach(() => {
+      jest.useRealTimers();
+    });
+
+    it('should end the session and call pc.close() because connection state is connecting', () => {
+      jest.useFakeTimers();
+      const parent = new FakeParent();
+      const session = new GenesysCloudMediaSession({
+        options: { parent },
+        sessionType: 'collaborateVideo',
+        allowIPv6: false
+      });
+      const reason = {
+        "reason": {
+          "condition": "success"
+        }
+      };
+      (session.pc.connectionState as any)  = 'connecting';
+      const spy = jest.spyOn(session, 'send').mockImplementation();
+      session.end('success');
+      jest.runAllTimers();
+
+      expect(spy).toHaveBeenCalledWith('session-terminate', reason);
+      expect(setTimeout).toHaveBeenCalledTimes(1);
+      expect(session.pc.close).toHaveBeenCalled();
+    });
+
+    it('should end the session - not call pc.close() because connection state is connected', () => {
+      jest.useFakeTimers();
+      const parent = new FakeParent();
+      const session = new GenesysCloudMediaSession({
+        options: { parent },
+        sessionType: 'collaborateVideo',
+        allowIPv6: false
+      });
+      const reason = {
+        "reason": {
+          "condition": "success"
+        }
+      };
+      (session.pc.connectionState as any)  = 'connected';
+      const spy = jest.spyOn(session, 'send').mockImplementation();
+      session.end('success');
+      jest.runAllTimers();
+
+      expect(spy).toHaveBeenCalledWith('session-terminate', reason);
+      expect(setTimeout).toHaveBeenCalledTimes(1);
+      expect(session.pc.close).toHaveBeenCalled();
+    });
+
+    it('should end the session - not call pc.close() because connection state is closed', () => {
+      jest.useFakeTimers();
+      const parent = new FakeParent();
+      const session = new GenesysCloudMediaSession({
+        options: { parent },
+        sessionType: 'collaborateVideo',
+        allowIPv6: false
+      });
+      const reason = {
+        "reason": {
+          "condition": "success"
+        }
+      };
+      (session.pc.connectionState as any)  = 'closed';
+      const spy = jest.spyOn(session, 'send').mockImplementation();
+      session.end('success');
+      jest.runAllTimers();
+
+      expect(spy).toHaveBeenCalledWith('session-terminate', reason);
+      expect(setTimeout).toHaveBeenCalledTimes(1);
+      expect(session.pc.close).not.toHaveBeenCalled();
     });
   });